TST: Changing tests to reflect new API changes.
